Create function for computer choice
Created and tested function that will randomly select the computers
choice for the rock, paper, scissors game. Confirmed that all three
responses can and do occur via the console. Probably could have written
the same thing with less code, but this made sure that my function was
only doing one thing, and not trying to assign the value and randomize
at the same time.