Add best of 7 game
Added code to run the game several times in a row and count wins and
losses. Chose not to use loops yet as we will learn about those and
my current knowledge is super limited as I haven't done them in a long
time. Tested everything out and it all works! The only thing I didn't
accomplish today was resetting the round in the event of a tie.