Create scoreboard and game over prompt
Created the scoreboard on html and wrote logic in script file to drive
the game and the score counting. Wrote functions to score each round
and check for the game over status of player or computer having 5
total wins. Wrote script to create elements to display a victory or
defeat message on the screen based on the result of the game and create
a button that will allow player to reset. Added boolean to check if the
game is still active based on the score and initialized it to true.